# PERF: MultiIndex.union