--- a/pandas/_libs/lib.pyx
+++ b/pandas/_libs/lib.pyx
@@ -312,46 +312,6 @@ def item_from_zerodim(val: object) -> object:
 
 @cython.wraparound(False)
 @cython.boundscheck(False)
-def fast_unique_multiple(ndarray left, ndarray right) -> list:
-    """
-    Generate a list indices we have to add to the left to get the union
-    of both arrays.
-
-    Parameters
-    ----------
-    left : np.ndarray
-        Left array that is used as base.
-    right : np.ndarray
-        right array that is checked for values that are not in left.
-        right can not have duplicates.
-
-    Returns
-    -------
-    list of indices that we have to add to the left array.
-    """
-    cdef:
-        Py_ssize_t j, n
-        list indices = []
-        set table = set()
-        object val, stub = 0
-
-    n = len(left)
-    for j in range(n):
-        val = left[j]
-        if val not in table:
-            table.add(val)
-
-    n = len(right)
-    for j in range(n):
-        val = right[j]
-        if val not in table:
-            indices.append(j)
-
-    return indices
-
-
-@cython.wraparound(False)
-@cython.boundscheck(False)
 def fast_unique_multiple_list(lists: list, sort: bool | None = True) -> list:
     cdef:
         list buf